refactor: move gc-helpers.js up to test/ , use it from other tests
Consolidate some copy/paste clones.
