+++ packages/SwingSet/test/stores/test-storeGC/test-refcount-management.js
@@ -32,7 +32,7 @@ import {
   validateStatusCheck,
   validateUpdate,
   validateWeakCheckEmpty,
-} from './gc-helpers.js';
+} from '../../gc-helpers.js';
 
 // These tests follow the model described in
 // ../virtualObjects/test-virtualObjectGC.js
