+++ packages/SwingSet/test/stores/test-storeGC/test-weak-key.js
@@ -30,7 +30,7 @@ import {
   validateStatusCheck,
   validateUpdate,
   validateWeakCheckEmpty,
-} from './gc-helpers.js';
+} from '../../gc-helpers.js';
 
 // These tests follow the model described in
 // ../virtualObjects/test-virtualObjectGC.js
