+++ packages/SwingSet/test/test-baggage.js
@@ -12,7 +12,7 @@ import {
   validateDone,
   validateReturned,
 } from './liveslots-helpers.js';
-import { capargs } from './util.js';
+import { validateCreateBuiltInTables } from './gc-helpers.js';
 
 function buildRootObject(vatPowers, vatParameters, baggage) {
   baggage.has('outside');
@@ -32,61 +32,6 @@ function stringVal(str) {
     body: JSON.stringify(str),
     slots: [],
   });
-}
-
-const stringSchema = JSON.stringify(
-  capargs([{ '@qclass': 'tagged', tag: 'match:kind', payload: 'string' }]),
-);
-
-const anyScalarSchema = JSON.stringify(
-  capargs([
-    {
-      '@qclass': 'tagged',
-      tag: 'match:scalar',
-      payload: { '@qclass': 'undefined' },
-    },
-  ]),
-);
-
-function validateCreateBuiltInTable(v, idx, idKey, schema, label) {
-  validate(v, matchVatstoreGet(idKey, NONE));
-  validate(v, matchVatstoreSet(`vc.${idx}.|nextOrdinal`, `1`));
-  validate(v, matchVatstoreSet(`vc.${idx}.|entryCount`, `0`));
-  validate(v, matchVatstoreSet(`vc.${idx}.|schemata`, schema));
-  validate(v, matchVatstoreSet(`vc.${idx}.|label`, label));
-  validate(v, matchVatstoreSet(idKey, `o+6/${idx}`));
-  validate(v, matchVatstoreGet(`vom.rc.o+6/${idx}`, NONE));
-  validate(v, matchVatstoreSet(`vom.rc.o+6/${idx}`, '1'));
-}
-
-function validateCreatePromiseWatcherKindTable(v, idx) {
-  validateCreateBuiltInTable(
-    v,
-    idx,
-    'watcherTableID',
-    anyScalarSchema,
-    'promiseWatcherByKind',
-  );
-}
-
-function validateCreateWatchedPromiseTable(v, idx) {
-  validateCreateBuiltInTable(
-    v,
-    idx,
-    'watchedPromiseTableID',
-    stringSchema,
-    'watchedPromises',
-  );
-}
-
-function validateCreateBaggage(v, idx) {
-  validateCreateBuiltInTable(v, idx, 'baggageID', stringSchema, 'baggage');
-}
-
-function validateCreateBuiltInTables(v) {
-  validateCreateBaggage(v, 1);
-  validateCreatePromiseWatcherKindTable(v, 2);
-  validateCreateWatchedPromiseTable(v, 3);
 }
 
 function validateSetup(v) {
