+++ packages/SwingSet/test/virtualObjects/test-virtualObjectGC.js
@@ -16,6 +16,7 @@ import {
   validateDone,
   validateReturned,
 } from '../liveslots-helpers.js';
+import { validateCreateBuiltInTables } from '../gc-helpers.js';
 import { capargs } from '../util.js';
 
 // Legs:
@@ -191,20 +192,6 @@ function facetRef(isf, vref, facet) {
 const cacheDisplacerVref = thingVref(false, 1);
 const fCacheDisplacerVref = thingVref(true, 1);
 const virtualHolderVref = `${holderBaseRef}/1`;
-
-const stringSchema = JSON.stringify(
-  capargs([{ '@qclass': 'tagged', tag: 'match:kind', payload: 'string' }]),
-);
-
-const anyScalarSchema = JSON.stringify(
-  capargs([
-    {
-      '@qclass': 'tagged',
-      tag: 'match:scalar',
-      payload: { '@qclass': 'undefined' },
-    },
-  ]),
-);
 
 function buildRootObject(vatPowers) {
   const { VatData, WeakMap, WeakSet } = vatPowers;
@@ -451,47 +438,6 @@ function validateFauxCacheDisplacerDeletion(v) {
   validateDelete(v, fCacheDisplacerVref);
   validateCheckNoWeakKeys(v, `${fCacheDisplacerVref}:0`);
   validateCheckNoWeakKeys(v, `${fCacheDisplacerVref}:1`);
-}
-
-function validateCreateBuiltInTable(v, idx, idKey, schema, label) {
-  validate(v, matchVatstoreGet(idKey, NONE));
-  validate(v, matchVatstoreSet(`vc.${idx}.|nextOrdinal`, `1`));
-  validate(v, matchVatstoreSet(`vc.${idx}.|entryCount`, `0`));
-  validate(v, matchVatstoreSet(`vc.${idx}.|schemata`, schema));
-  validate(v, matchVatstoreSet(`vc.${idx}.|label`, label));
-  validate(v, matchVatstoreSet(idKey, `o+6/${idx}`));
-  validate(v, matchVatstoreGet(`vom.rc.o+6/${idx}`, NONE));
-  validate(v, matchVatstoreSet(`vom.rc.o+6/${idx}`, '1'));
-}
-
-function validateCreatePromiseWatcherKindTable(v, idx) {
-  validateCreateBuiltInTable(
-    v,
-    idx,
-    'watcherTableID',
-    anyScalarSchema,
-    'promiseWatcherByKind',
-  );
-}
-
-function validateCreateWatchedPromiseTable(v, idx) {
-  validateCreateBuiltInTable(
-    v,
-    idx,
-    'watchedPromiseTableID',
-    stringSchema,
-    'watchedPromises',
-  );
-}
-
-function validateCreateBaggage(v, idx) {
-  validateCreateBuiltInTable(v, idx, 'baggageID', stringSchema, 'baggage');
-}
-
-function validateCreateBuiltInTables(v) {
-  validateCreateBaggage(v, 1);
-  validateCreatePromiseWatcherKindTable(v, 2);
-  validateCreateWatchedPromiseTable(v, 3);
 }
 
 function validateKindMetadata(v, kindID, tag) {
